app/requirements.txt
app==0.0.1
fastapi==0.115.12
openai==1.73.0
Pillow==11.2.1
pydantic==2.11.3
python-dotenv==1.1.0
python-multipart==0.0.7
numpy==1.24.3
torch==2.2.0
transformers==4.38.0
accelerate==0.28.0
uvicorn==0.34.1
